+++ src/hardis-status-provider.ts
@@ -16,6 +16,7 @@ import {
 import { Logger } from "./logger";
 import simpleGit from "simple-git";
 import { ThemeUtils } from "./themeUtils";
+import { getConfig } from "./utils/pipeline/sfdxHardisConfig";
 
 export class HardisStatusProvider
   implements vscode.TreeDataProvider<StatusTreeItem>
@@ -112,9 +113,11 @@ export class HardisStatusProvider
     let orgDisplayCommand = "sf org display";
     if (options.devHub) {
       const devHubAliasCommand = "sf config get target-dev-hub";
-      const devHubAliasRes = await execSfdxJson(devHubAliasCommand, this, {
+      const devHubAliasRes = await execSfdxJson(devHubAliasCommand, {
         fail: false,
         output: false,
+        cacheSection: "project",
+        cacheExpiration: 1000 * 60 * 60, // 1 hour
       });
       if (
         devHubAliasRes &&
@@ -123,7 +126,7 @@ export class HardisStatusProvider
         devHubAliasRes.result[0].value
       ) {
         devHubUsername = devHubAliasRes.result[0].value;
-        orgDisplayCommand += ` -- ${devHubUsername}`;
+        orgDisplayCommand += ` --target-org ${devHubUsername}`;
       } else {
         items.push({
           id: "org-not-connected-devhub",
@@ -135,9 +138,11 @@ export class HardisStatusProvider
         return items;
       }
     }
-    const orgInfoResult = await execSfdxJson(orgDisplayCommand, this, {
+    const orgInfoResult = await execSfdxJson(orgDisplayCommand,  {
       fail: false,
       output: false,
+      cacheSection: orgDisplayCommand.includes("--target-org") ? "orgs": "project",
+      cacheExpiration: orgDisplayCommand.includes("--target-org") ? 1000 * 60 * 60 * 24 * 90 : 1000 * 60 * 15, // 90 days for named orgs, 15 minutes for default org
     });
     if (orgInfoResult.result || orgInfoResult.id) {
       const orgInfo = orgInfoResult.result || orgInfoResult;
@@ -232,7 +237,6 @@ Maybe update sourceApiVersion in your sfdx-project.json ? (but be careful if you
         if (config?.poolConfig) {
           const poolViewRes = await execSfdxJson(
             "sf hardis:scratch:pool:view",
-            this,
             { output: false, fail: false },
           );
           if (
@@ -414,16 +418,10 @@ Note: Disable disableGitMergeRequiredCheck in settings to skip this check.`;
           tooltip: gitTooltip,
           command: gitCommand,
         });
-        if (isCachePreloaded()) {
-          // Merge request info
-          const mergeRequestRes = await execSfdxJson(
-            "sf hardis:config:get --level user",
-            this,
-            { fail: false, output: true },
-          );
-          if (mergeRequestRes?.result?.config?.mergeRequests) {
+          const userConfig = await getConfig("user");
+          if (userConfig.mergeRequests) {
             const mergeRequests =
-              mergeRequestRes.result.config.mergeRequests.filter(
+              userConfig.mergeRequests.filter(
                 (mr: any) =>
                   mr !== null &&
                   mr.branch === currentBranch &&
@@ -459,7 +457,6 @@ Note: Disable disableGitMergeRequiredCheck in settings to skip this check.`;
               });
             }
           }
-        }
       }
     }
     return items;