+++ src/webviews/lwc-ui/modules/s/pipeline/pipeline.js
@@ -3,6 +3,7 @@
 // @ts-nocheck
 // eslint-env es6
 import { LightningElement, api, track } from "lwc";
+import 's/forceLightTheme'; // Ensure light theme is applied
 
 export default class Pipeline extends LightningElement {
   @track prButtonInfo;