+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## [Unreleased]
 
+## [6.0.10 (pre-release)] 2024-08-17
+
 - Add debug options for sfdx-hardis commands
 - Add new sandbox refresh commands
 - New UI to update VsCode SFDX-Hardis configuration
 - Fix issue on MacOS with background commands
+- Use VsCode Cache manager to improve performances
+- Add Pipeline Settings command & LWC
+- Build custom LWC helptext component to handle multiline display
 
 ## [6.0.9 (pre-release)] 2024-08-14